refactored code to segregate components
