+++ 07_uns_graphql/src/uns_graphql/subscriptions/mqtt.py
@@ -1,0 +1,63 @@
+"""
+This module contains subscription methods fro GraphQL
+It provides asynchronous generators for subscribing to MQTT topics.
+"""
+
+import asyncio
+import logging
+import random
+import time
+import typing
+
+import strawberry
+from aiomqtt import Client, MqttError
+
+from uns_graphql.graphql_config import MQTTConfig
+from uns_graphql.input.mqtt import MQTTTopicInput
+from uns_graphql.type.mqtt_event import MQTTMessage
+
+LOGGER = logging.getLogger(__name__)
+
+
+@strawberry.type
+class MQTTSubscription:
+    """
+    Subscription class providing methods for subscribing to MQTT messages.
+    """
+
+    @strawberry.subscription(
+        description=" Subscribe to MQTT messages based on provided list of topics. MQTT wildcards are supported "
+    )
+    async def get_mqtt_messages(self, topics: list[MQTTTopicInput]) -> typing.AsyncGenerator[MQTTMessage, None]:
+        """
+        Subscribe to MQTT messages based on provided topics.
+
+        Args:
+            topics (List of MQTTTopicInput): List of Topics to subscribe to. Supports wildcards
+
+        Yields:
+            typing.AsyncGenerator[UNSMessage, None]: Asynchronously generates UNS event messages
+        """
+        try:
+            client_id = f"graphql-{time.time()}-{random.randint(0, 1000)}"  # noqa: S311
+            async with Client(
+                client_id=client_id,
+                clean_session=MQTTConfig.clean_session,
+                protocol=MQTTConfig.version,
+                transport=MQTTConfig.transport,
+                hostname=MQTTConfig.host,
+                port=MQTTConfig.port,
+                username=MQTTConfig.username,
+                password=MQTTConfig.password,
+                keepalive=MQTTConfig.keep_alive,
+                tls_params=MQTTConfig.tls_params,
+                tls_insecure=MQTTConfig.tls_insecure,
+            ) as client:
+                async with client.messages() as messages:
+                    for mqtt_topic in topics:
+                        await client.subscribe(topic=mqtt_topic.topic, qos=MQTTConfig.qos, properties=MQTTConfig.properties)
+                    async for msg in messages:
+                        yield MQTTMessage(topic=str(msg.topic), payload=msg.payload)
+        except MqttError as ex:
+            LOGGER.error("Error while connecting to MQTT Broker %s", str(ex), stack_info=True, exc_info=True)
+            await asyncio.sleep(MQTTConfig.retry_interval)
