+++ 07_uns_graphql/test/types/test_historical_event.py
@@ -1,3 +1,5 @@
+from datetime import datetime
+
 import pytest
 import strawberry
 from uns_graphql.type.basetype import JSONPayload
@@ -7,7 +9,10 @@ from uns_graphql.type.historical_event import HistoricalUNSEvent
 @pytest.fixture
 def sample_historical_event():
     return HistoricalUNSEvent(
-        publisher="Publisher", timestamp=1234567890, topic="sample/topic", payload=JSONPayload(data={"key": "value"})
+        publisher="Publisher",
+        timestamp=datetime.fromtimestamp(1234567890),
+        topic="sample/topic",
+        payload=JSONPayload(data={"key": "value"}),
     )
 
 
@@ -20,7 +25,7 @@ def test_historical_event_published_by(sample_historical_event):
 
 
 def test_historical_event_timestamp(sample_historical_event):
-    assert sample_historical_event.timestamp == 1234567890
+    assert sample_historical_event.timestamp == datetime.fromtimestamp(1234567890)
 
 
 def test_historical_event_topic(sample_historical_event):
@@ -30,17 +35,6 @@ def test_historical_event_topic(sample_historical_event):
 def test_historical_event_payload(sample_historical_event):
     assert isinstance(sample_historical_event.payload, JSONPayload)
     assert sample_historical_event.payload.data == '{"key": "value"}'
-
-
-# data validations not implemented on types returning data from back end
-# def test_historical_event_invalid_payload():
-#     with pytest.raises(TypeError):
-#         HistoricalUNSEvent(
-#             published_by="Publisher",
-#             timestamp=1234567890,
-#             topic="sample/topic",
-#             payload={"invalid": "payload"},  # Providing an invalid payload
-#         )
 
 
 def test_strawberry_type(sample_historical_event):
